feat: added layout package BREAKING CHANGE: header package removed BREAKING CHANGE: footer package removed
