--- packages/footer/CHANGELOG.md
@@ -1,52 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [0.1.12](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.11...@sk-web-gui/footer@0.1.12) (2023-10-10)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## 0.1.11 (2023-05-16)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.10](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.9...@sk-web-gui/footer@0.1.10) (2023-04-21)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.9](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.8...@sk-web-gui/footer@0.1.9) (2023-04-19)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.8](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.7...@sk-web-gui/footer@0.1.8) (2023-02-14)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.7](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.6...@sk-web-gui/footer@0.1.7) (2022-11-22)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.6](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.5...@sk-web-gui/footer@0.1.6) (2022-11-07)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.5](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.4...@sk-web-gui/footer@0.1.5) (2022-11-07)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.4](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.3...@sk-web-gui/footer@0.1.4) (2022-11-07)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.3](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.2...@sk-web-gui/footer@0.1.3) (2022-11-04)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## [0.1.2](https://github.com/Sundsvallskommun/web-shared-components/compare/@sk-web-gui/footer@0.1.1...@sk-web-gui/footer@0.1.2) (2022-09-22)
-
-**Note:** Version bump only for package @sk-web-gui/footer
-
-## 0.1.1 (2022-09-05)
-
-**Note:** Version bump only for package @sk-web-gui/footer
