--- packages/footer/README.md
@@ -1,5 +1,0 @@
-## Installation
-
-```sh
-yarn add @sk-web-gui/footer
-```
