+++ packages/layout/src/header/assets/logo.tsx
@@ -1,0 +1,8 @@
+export const Logo = () => (
+  <svg width="33.5" height="56" viewBox="0 0 43 72" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      fill="currentColor"
+      d="M10.9192 7.9255C10.9192 7.9255 9.43796 10.5325 12.0449 11.2435C15.3629 12.073 18.9179 12.6062 22.2359 13.9097C23.1839 14.2652 22.1766 15.154 21.5249 15.0947C15.7777 13.495 10.3267 11.5397 3.75 11.836C5.23124 13.6135 7.42348 16.0427 7.89747 18.4719C10.2082 16.6352 13.7632 15.6279 16.5479 17.1092L13.2299 18.2942C8.01597 20.2494 4.461 25.7596 3.92775 31.0921C3.45375 37.0171 6.59398 42.1718 11.808 44.838C18.0292 48.156 28.0423 46.734 30.8863 53.37C31.5381 54.9697 31.5381 58.1099 30.1753 60.0059C27.8053 63.0277 24.4281 63.9164 20.4584 63.4424C20.4584 63.4424 15.7777 62.7907 13.0522 60.7169C9.02322 57.6952 7.89747 53.6662 7.36423 49.1632H4.99424C5.58674 54.8512 4.99424 60.5984 4.224 65.9901L6.23848 66.2271C6.23848 66.2271 6.77173 64.8051 7.36423 63.9757C7.83822 63.5017 8.60847 63.3832 9.20097 63.7979C16.1924 68.5971 27.5683 68.5971 34.1451 64.1534C38.648 60.8354 40.307 55.7992 38.9443 50.526C35.9226 39.9203 23.8356 41.1053 15.9554 37.4318C13.3484 36.4246 10.6822 33.8768 10.6822 30.7959C10.623 27.5964 11.9264 25.0486 14.5334 23.2119C15.5999 22.5602 16.7257 22.4417 16.7257 22.4417C14.5927 24.6339 13.2892 27.1816 13.6447 30.3811C13.9409 32.9881 16.1924 35.4766 18.6809 36.2468C21.1101 36.8986 23.3024 35.9506 25.1391 34.5286L24.9021 33.4028C23.8356 33.6991 22.4729 33.6991 21.4064 33.4028C19.2142 32.6326 18.3847 30.4404 18.6217 28.2481C18.7994 27.3001 19.5697 26.6484 20.6361 26.9446C26.0279 29.1961 25.0206 36.3653 30.1161 38.8538C31.8343 39.4463 33.3748 38.9131 34.4413 38.1428C37.1075 36.2468 38.0555 33.8176 38.7665 30.7959C37.3445 32.3956 36.041 34.1731 33.9081 34.5286C32.8416 34.7656 32.3083 34.5286 31.7158 33.6398C29.6421 30.3219 28.9903 24.8116 24.0726 24.2784C23.1246 24.2784 22.8876 24.2191 21.1101 24.4561C22.3544 23.4489 23.8356 22.4417 25.4946 22.3824C27.9238 22.3232 30.4716 23.9821 31.4196 25.6411C31.8936 26.5891 31.7751 28.0704 31.0641 28.9591L32.1898 29.7886C34.0266 27.4186 37.9963 24.3969 37.937 22.3824C37.6408 18.4719 31.7158 16.1019 28.2201 14.7392C28.2201 14.7392 28.8126 13.1987 28.5163 12.073C27.8646 10.5325 26.3834 10.2955 25.0206 9.88073C22.4729 9.16974 18.3847 8.81424 16.4887 7.15525C14.5927 5.49626 15.4814 3.71877 15.9554 1.82278C11.1562 7.27375 10.9192 7.9255 10.9192 7.9255ZM26.2649 16.1612C26.8574 16.4574 27.4499 17.0499 27.7461 17.6424L22.4136 16.7537C23.4801 15.9834 25.0206 15.391 26.2649 16.1612Z"
+    />
+  </svg>
+);
