+++ packages/react/src/index.ts
@@ -7,12 +7,11 @@ export * from '@sk-web-gui/card';
 export * from '@sk-web-gui/cookie-consent';
 export * from '@sk-web-gui/divider';
 export * from '@sk-web-gui/filter';
-export * from '@sk-web-gui/footer';
 export * from '@sk-web-gui/forms';
-export * from '@sk-web-gui/header';
 export * from '@sk-web-gui/icon';
 export * from '@sk-web-gui/image';
 export * from '@sk-web-gui/label';
+export * from '@sk-web-gui/layout';
 export * from '@sk-web-gui/link';
 export * from '@sk-web-gui/menu-vertical';
 export * from '@sk-web-gui/menubar';
